Restructure package files and fix string references.
