+++ ivy/functional/backends/jax/__init__.py
@@ -33,10 +33,10 @@ if not ivy.is_local():
 if not ivy.is_local():
     _module_in_memory = sys.modules[__name__]
 else:
-    global_backend_compiler = sys.modules["ivy.backend_compiler"]
+    global_backend_compiler = sys.modules["ivy.utils.backend.backend_compiler"]
     _module_in_memory = global_backend_compiler.IMPORT_CACHE[__name__]
 
-use = ivy.backend_handler.ContextManager(_module_in_memory)
+use = ivy.utils.backend.ContextManager(_module_in_memory)
 
 # noinspection PyUnresolvedReferences
 JaxArray = Union[
