+++ ivy/functional/backends/tensorflow/__init__.py
@@ -15,10 +15,10 @@ backend_version = {"version": tf.__version__}
 if not ivy.is_local():
     _module_in_memory = sys.modules[__name__]
 else:
-    global_backend_compiler = sys.modules["ivy.backend_compiler"]
+    global_backend_compiler = sys.modules["ivy.utils.backend.backend_compiler"]
     _module_in_memory = global_backend_compiler.IMPORT_CACHE[__name__]
 
-use = ivy.backend_handler.ContextManager(_module_in_memory)
+use = ivy.utils.backend.ContextManager(_module_in_memory)
 
 NativeArray = Tensor
 NativeVariable = Tensor
