+++ ivy/functional/ivy/activations.py
@@ -5,7 +5,7 @@ import sys
 
 # local
 import ivy
-from ivy.backend_handler import current_backend
+from ivy.utils.backend import current_backend
 from ivy.func_wrapper import (
     handle_array_function,
     handle_out_argument,
