+++ ivy/functional/ivy/control_flow_ops.py
@@ -1,4 +1,4 @@
-from ivy.backend_handler import current_backend
+from ivy.utils.backend import current_backend
 from ivy.func_wrapper import (
     handle_array_like_without_promotion,
     to_native_arrays_and_back,
