+++ ivy/functional/ivy/creation.py
@@ -8,7 +8,7 @@ import numpy as np
 # local
 import ivy
 from ivy import to_ivy
-from ivy.backend_handler import current_backend
+from ivy.utils.backend import current_backend
 from ivy.exceptions import handle_exceptions
 from ivy.func_wrapper import (
     handle_array_function,
