+++ ivy/functional/ivy/experimental/activations.py
@@ -3,7 +3,7 @@ from typing import Union, Optional
 
 # local
 import ivy
-from ivy.backend_handler import current_backend
+from ivy.utils.backend import current_backend
 from ivy.exceptions import handle_exceptions
 from ivy.func_wrapper import (
     handle_nestable,
