+++ ivy/functional/ivy/experimental/creation.py
@@ -5,7 +5,7 @@ from typing import Union, Tuple, Optional
 
 # local
 import ivy
-from ivy.backend_handler import current_backend
+from ivy.utils.backend import current_backend
 from ivy.exceptions import handle_exceptions
 from ivy.func_wrapper import (
     infer_device,
