+++ ivy/functional/ivy/experimental/norms.py
@@ -2,7 +2,7 @@ from typing import Union, Optional
 
 # local
 import ivy
-from ivy.backend_handler import current_backend
+from ivy.utils.backend import current_backend
 from ivy.func_wrapper import (
     to_native_arrays_and_back,
     handle_out_argument,
