+++ ivy/functional/ivy/linear_algebra.py
@@ -3,7 +3,7 @@ from typing import Union, Optional, Tuple, Literal, List, Sequence
 
 # local
 import ivy
-from ivy.backend_handler import current_backend
+from ivy.utils.backend import current_backend
 from ivy.func_wrapper import (
     handle_array_function,
     to_native_arrays_and_back,
