+++ ivy/functional/ivy/manipulation.py
@@ -6,7 +6,7 @@ from numpy.core.numeric import normalize_axis_tuple
 
 # local
 import ivy
-from ivy.backend_handler import current_backend
+from ivy.utils.backend import current_backend
 from ivy.func_wrapper import (
     handle_array_function,
     to_native_arrays_and_back,
