+++ ivy/utils/backend/__init__.py
@@ -1,0 +1,2 @@
+from . import backend_handler
+from .backend_handler import *
