+++ ivy/utils/backend/backend_compiler.py
@@ -164,7 +164,7 @@ class ImportTransformer(ast.NodeTransformer):
             tree.body.insert(
                 self.insert_index,
                 ast.ImportFrom(
-                    module="ivy.backend_compiler",
+                    module="ivy.utils.backend.backend_compiler",
                     names=[ast.alias(name="_ivy_fromimport")],
                     level=0,
                 ),
@@ -172,7 +172,7 @@ class ImportTransformer(ast.NodeTransformer):
             tree.body.insert(
                 self.insert_index,
                 ast.ImportFrom(
-                    module="ivy.backend_compiler",
+                    module="ivy.utils.backend.backend_compiler",
                     names=[ast.alias(name="_ivy_absolute_import")],
                     level=0,
                 ),
@@ -350,14 +350,18 @@ def with_backend(backend: str):
     sys.meta_path.insert(0, FINDER)
     ivy_pack = _ivy_import_module("ivy")
     ivy_pack._is_local = True
-    backend_module = _ivy_import_module(ivy_pack.backend_handler._backend_dict[backend], ivy_pack.__package__)
+    backend_module = _ivy_import_module(
+        ivy_pack.backend_handler._backend_dict[backend], ivy_pack.__package__
+    )
     # TODO temporary
     if backend == "numpy":
         ivy_pack.set_default_device("cpu")
     elif backend == "jax":
         ivy_pack.set_global_attr("RNG", ivy_pack.functional.backends.jax.random.RNG)
     # We know for sure that the backend stack is empty, no need to do backend unsetting
-    ivy_pack.backend_handler._set_backend_as_ivy(ivy_pack.__dict__.copy(), ivy_pack, backend_module)
+    ivy_pack.backend_handler._set_backend_as_ivy(
+        ivy_pack.__dict__.copy(), ivy_pack, backend_module
+    )
     # Remove access to specific modules on local Ivy
     for module in modules_to_remove:
         for fn in inspect.getmembers(ivy_pack.__dict__[module], inspect.isfunction):
